+++ packages/cortex-core/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1-beta.2](https://github.com/tecsinapse/design-system/compare/@tecsinapse/cortex-core@0.3.1-beta.1...@tecsinapse/cortex-core@0.3.1-beta.2) (2024-08-21)
+
+**Note:** Version bump only for package @tecsinapse/cortex-core
+
+## [0.3.1-beta.1](https://github.com/tecsinapse/design-system/compare/@tecsinapse/cortex-core@0.3.1-beta.0...@tecsinapse/cortex-core@0.3.1-beta.1) (2024-08-21)
+
+**Note:** Version bump only for package @tecsinapse/cortex-core
+
+## [0.3.1-beta.0](https://github.com/tecsinapse/design-system/compare/@tecsinapse/cortex-core@0.3.0...@tecsinapse/cortex-core@0.3.1-beta.0) (2024-08-20)
+
+**Note:** Version bump only for package @tecsinapse/cortex-core
+
 # [0.3.0](https://github.com/tecsinapse/design-system/compare/@tecsinapse/cortex-core@0.3.0-beta.14...@tecsinapse/cortex-core@0.3.0) (2024-08-06)
 
 **Note:** Version bump only for package @tecsinapse/cortex-core